Created Projections folder and moved Projection file into it. Synced namespaces and usings
