+++ src/Bruhunter.WebApp/Pages/Candidates.razor
@@ -1,6 +1,7 @@
 ﻿@page "/candidates"
 @using System.Net.Http.Headers
 @using Bruhunter.WebApp.Pages
+@using Bruhunter.Shared.Projections
 @namespace Bruhunter.Shared.Documents
 
 @inject HttpClient Http
