+++ src/Bruhunter.WebApp/Pages/EditCandidateModal.razor
@@ -1,5 +1,6 @@
 ﻿@using Bruhunter.Shared.Documents
 @using System.Net.Http.Headers
+@using Bruhunter.Shared.Projections
 
 @inject HttpClient Http
